backend/requirements.txt
alembic==1.0.0
bcrypt==3.1.4
certifi==2018.8.24
flask-admin==1.5.6
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.1.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Pillow==6.2.1
pymongo==3.7.1
urllib3==1.23
requests==2.19.1
regex==2018.8.29
Werkzeug==0.14.1
WTForms==2.2.1